Update the project setup and integrate DevSupport
This commit introduces several changes to the set up of the project and makes it compatible with the @react-buddy/ide-toolbox package. Changes include, updates to rules in .eslintrc.cjs file to manage linter preferences and code style. Packages - prettier and @react-buddy/ide-toolbox are added to the project dependencies. New `DevSupport` component is included in the main.jsx to allow loading of dedicated development components. Also, multiple new files are added in src/dev directory to facilitate the use of dev tooling during development, which includes previews, palette, and hooks.