Refactor CSS and improve navigation system
This commit removes unnecessary CSS and standardises the styles for multiple components, affecting the html, body, .app, .content, and #root classes. The typography setting was added to the theme for uniformity in src/style/theme.js.

In src/components/layout/Layout2.jsx, the navigation system was enhanced by adding the Link component from react-router-dom. This allows for the selection of navigation items to link to specific routes.

Finally, different font weights for Roboto were imported in the main module to maintain visual consistency across the application.