Rename "About" to "Ayuda" and enhance UI UX for sidebar menu
The 'About' page was renamed to 'Ayuda' to localize the application for Spanish-speaking users. This change ensures a more user-friendly experience for this demographic. The 'Clientes' page was updated with more content and imported into the main layout. Additionally, changes were made to enhance the UI and UX of the sidebar menu in 'Layout2.jsx'. The 'AppRoutes.jsx' was also updated to reflect these changes. This included changing the order of routes and replacing 'About.jsx' with 'Ayuda.jsx'.