Update UI elements and dependencies
Removed explicit border styling from Stock and Productos components and changed the text color in HeaderTabla component to improve clarity. Added a placeholder image to add variety to the interface.

Improvements to the web application also include the addition of a new theme switcher in ThemeSwitch.js and adding various charts in Chart.js to display data more effectively.

Severals changes to dependencies have been made; Addition of "recharts" dependency in package.json and package-lock.json. Multiple other files were added and updates for the improved functioning of the application.

This commit enhances the UI and UX of the platform.